+++ modules/usb/src/drivers/usb_drivers.c
@@ -85,7 +85,7 @@ int usb_drivers_probe(
       const usb_device_t* pdevice,
       usb_driver_handle_t offset,
       const uint8_t*      buffer,
-      uint8_t             length
+      uint16_t            length
 )
 {
    int ret;
@@ -127,7 +127,7 @@ int usb_drivers_assign(
       usb_stack_t*        pstack,
       uint16_t            id,
       const uint8_t*      buffer,
-      uint8_t             length,
+      uint16_t            length,
       usb_driver_handle_t handle
 )
 {
